requirements.txt
Django==3.0.4
psycopg2==2.8.4
Pillow==6.2.1
